Refactor removing duplicate code
I removed some code in the Defender class that was repeating itself and
made a method for it. The code where I removed was than calling the
method instead of having repetitive code.
I also removed a bunch of scanner objects in the Manager class because I
kept doing new Scanner() everytime I asked for the user input, instead I
did it once and kept reusing it.